jdbc/src/main/java/org/springframework/transaction/support/TransactionSynchronizationManager.java
package org.springframework.transaction.support;

import java.util.HashMap;
import javax.annotation.Nullable;
import javax.sql.DataSource;
import java.sql.Connection;
import java.util.Map;

public abstract class TransactionSynchronizationManager {

    private static final ThreadLocal<Map<DataSource, Connection>> resources = ThreadLocal.withInitial(HashMap::new);

    private TransactionSynchronizationManager() {}

    @Nullable
    public static Connection getResource(DataSource key) {
        return resources.get().getOrDefault(key, null);
    }

    public static void bindResource(DataSource key, Connection value) {
        resources.get().put(key, value);
    }

    public static Connection unbindResource(DataSource key) {
        return resources.get().remove(key);
    }
}
